+++ modules/apps/layout/layout-page-template-admin-web/src/main/java/com/liferay/layout/page/template/admin/web/internal/headless/delivery/dto/v1_0/structure/importer/helper/PortletConfigurationImporterHelper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.headless.delivery.dto.v1_0.structure.importer.util;
+package com.liferay.layout.page.template.admin.web.internal.headless.delivery.dto.v1_0.structure.importer.helper;
 
 import com.liferay.layout.page.template.admin.web.internal.importer.PortletConfigurationImporterTracker;
 import com.liferay.layout.page.template.importer.PortletConfigurationImporter;
